install: Move command logic into cli package
