+++ cli/install/install.go
@@ -14,7 +14,7 @@
 //
 // SPDX-License-Identifier: Apache-2.0
 
-package cli
+package install
 
 import (
 	"fmt"
